testing

+++ index.php
@@ -50,26 +50,27 @@ if(isset($_POST) and $_SERVER['REQUEST_METHOD'] == "POST"){
 				continue; // Skip invalid file formats
 			}
 	        else{ // No error found! Move uploaded files 
-	        	echo $_FILES['files']['name'][$f];
-	        	if (move_uploaded_file($_FILES['files']['tmp_name'][$f], "uploads/".$name))
-				{
-					$file = fopen("uploads/".$name, "rb");
-					$contents = fread($file, filesize("uploads/".$name));
-					fclose($file);
-					$fileImage = ParseFile::createFromData($contents, $name);
-					$fileImage->save();
+	        	if (!file_exists("uploads/".$name)) {
+	        		if (move_uploaded_file($_FILES['files']['tmp_name'][$f], "uploads/".$name))
+					{
+						$file = fopen("uploads/".$name, "rb");
+						$contents = fread($file, filesize("uploads/".$name));
+						fclose($file);
+						$fileImage = ParseFile::createFromData($contents, $name);
+						$fileImage->save();
 
-					$images = new ParseObject("Photos");
-					$images->set("FileName", $name);
-					$images->set("FileImage", $fileImage);
-					try {
-					  $images->save();
-					} catch (ParseException $ex) {  
-					  // Execute any logic that should take place if the save fails.
-					  // error is a ParseException object with an error code and message.
-					  echo 'There is a problem!: ' . $ex->getMessage();
+						$images = new ParseObject("Photos");
+						$images->set("FileName", $name);
+						$images->set("FileImage", $fileImage);
+						try {
+						  $images->save();
+						} catch (ParseException $ex) {  
+						  // Execute any logic that should take place if the save fails.
+						  // error is a ParseException object with an error code and message.
+						  echo 'There is a problem!: ' . $ex->getMessage();
+						}
 					}
-				}
+	        	}
 	        }
 	    }
 	}
